WIP - Refactor Editor.ts into smaller source files
* Created FormattingService and ListService
* Moved functions from Editor into the services
* Partitioned test files to test the respective services

Jira ticket: CAMS-526
